Moved enqueueFillingResultItemsFromIDs
